# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API OnBoardPro",
    "version": "1.0.0",
    "description": "Documentación de la app de OnBoardPro"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/courses/get/all": {
      "get": {
        "summary": "Obtiene todos los cursos",
        "tags": [
          "Cursos"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "buscar",
            "description": "Coincidencia parcial en el nombre del curso.",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "estado",
            "schema": {
              "type": "string",
              "enum": [
                "Abierto",
                "Cerrado"
              ]
            }
          },
          {
            "in": "query",
            "name": "pagina",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "porPagina",
            "schema": {
              "type": "integer",
              "default": 50,
              "maximum": 200
            }
          }
        ],
        "description": "El cuerpo es el array de siempre. La paginación viaja en cabeceras: `X-Total-Count`, `X-Page`, `X-Per-Page` y `X-Total-Pages`. Un instructor recibe únicamente los cursos que tiene asignados; el filtro es automático y no se puede desactivar.\n",
        "responses": {
          "200": {
            "description": "Lista de todos los cursos"
          }
        }
      }
    },
    "/api/courses/get/id/{id}": {
      "get": {
        "summary": "Obtiene un curso por su ID",
        "tags": [
          "Cursos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID del curso"
          }
        ],
        "responses": {
          "200": {
            "description": "Curso encontrado"
          },
          "404": {
            "description": "Curso no encontrado"
          }
        }
      }
    },
    "/api/courses/create": {
      "post": {
        "summary": "Crea un nuevo curso",
        "tags": [
          "Cursos"
        ],
        "description": "El curso queda en la empresa del usuario autenticado; `tenantId` se toma del token y se ignora si viene en el cuerpo. El `id` lo genera el servidor. La portada se sube después con POST /api/courses/{id}/banner.\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "instructor"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "maxLength": 50,
                    "example": "Inducción para nuevos ingresos"
                  },
                  "instructor": {
                    "type": "string",
                    "format": "email",
                    "description": "Correo del instructor a cargo.",
                    "example": "instructor@empresa.com"
                  },
                  "grupo": {
                    "type": "string",
                    "description": "Id del equipo al que se asigna el curso.",
                    "example": "grp-uc-a"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "Abierto",
                      "Cerrado"
                    ],
                    "default": "Abierto"
                  },
                  "actividades": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Ids de actividades ya creadas."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Curso creado correctamente"
          },
          "400": {
            "description": "Error en los datos o validación fallida"
          }
        }
      }
    },
    "/api/courses/me": {
      "get": {
        "summary": "Obtiene los cursos del usuario autenticado",
        "tags": [
          "Cursos"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de cursos asociados al usuario autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "abc123"
                      },
                      "name": {
                        "type": "string",
                        "example": "Introducción a la programación"
                      },
                      "description": {
                        "type": "string",
                        "example": "Curso básico de fundamentos de programación"
                      },
                      "image": {
                        "type": "string",
                        "example": "https://example.com/course-image.jpg"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado. El token JWT es inválido o no fue enviado"
          },
          "500": {
            "description": "Error interno del servidor"
          }
        }
      }
    },
    "/api/courses/{id}/banner": {
      "post": {
        "summary": "Sube o reemplaza la portada de un curso",
        "tags": [
          "Cursos"
        ],
        "description": "Recibe una imagen en multipart/form-data bajo el campo `banner`. Máximo 5 MB, solo JPEG, PNG, WebP o GIF. Queda en el bucket público \"banners\" de Supabase Storage y la portada anterior se borra.\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "banner": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Portada actualizada, devuelve bannerUrl"
          },
          "400": {
            "description": "Sin archivo, tipo no permitido o supera los 5 MB"
          },
          "404": {
            "description": "El curso no existe o es de otra empresa"
          }
        }
      }
    },
...
  },
...
}