# routes/coursesRoutes.js
const express = require("express");
const router = express.Router();
const CourseController = require("../controllers/courseController");
const verifyToken = require("../middlewares/jwt.js");
const requireRole = require("../middlewares/requireRole.js");
const subida = require("../middlewares/subida.js");

/**
 * @swagger
 * tags:
 *   name: Cursos
 *   description: Endpoints relacionados con cursos y sus usuarios
 */

/**
 * @swagger
 * /api/courses/get/all:
 *   get:
 *     summary: Obtiene todos los cursos
 *     tags: [Cursos]
 *     security:
 *       - bearerAuth: []
 *     parameters:
 *       - in: query
 *         name: buscar
 *         description: Coincidencia parcial en el nombre del curso.
 *         schema: { type: string }
 *       - in: query
 *         name: estado
 *         schema: { type: string, enum: [Abierto, Cerrado] }
 *       - in: query
 *         name: pagina
 *         schema: { type: integer, default: 1 }
 *       - in: query
 *         name: porPagina
 *         schema: { type: integer, default: 50, maximum: 200 }
 *     description: >
 *       El cuerpo es el array de siempre. La paginación viaja en cabeceras:
 *       `X-Total-Count`, `X-Page`, `X-Per-Page` y `X-Total-Pages`.
 *       Un instructor recibe únicamente los cursos que tiene asignados;
 *       el filtro es automático y no se puede desactivar.
 *     responses:
 *       200:
 *         description: Lista de todos los cursos
 */
router.get("/get/all", verifyToken, CourseController.getAll);

/**
 * @swagger
 * /api/courses/get/id/{id}:
 *   get:
 *     summary: Obtiene un curso por su ID
 *     tags: [Cursos]
 *     parameters:
 *       - in: path
 *         name: id
 *         required: true
 *         schema:
 *           type: string
 *         description: ID del curso
 *     responses:
 *       200:
 *         description: Curso encontrado
 *       404:
 *         description: Curso no encontrado
 */
router.get("/get/id/:id", verifyToken, CourseController.getById);

/**
 * @swagger
 * /api/courses/create:
 *   post:
 *     summary: Crea un nuevo curso
 *     tags: [Cursos]
 *     description: >
 *       El curso queda en la empresa del usuario autenticado; `tenantId` se toma
 *       del token y se ignora si viene en el cuerpo. El `id` lo genera el
 *       servidor. La portada se sube después con POST /api/courses/{id}/banner.
 *     security:
 *       - bearerAuth: []
 *     requestBody:
 *       required: true
 *       content:
 *         application/json:
 *           schema:
 *             type: object
 *             required:
 *               - name
 *               - instructor
 *             properties:
 *               name:
 *                 type: string
 *                 maxLength: 50
 *                 example: Inducción para nuevos ingresos
 *               instructor:
 *                 type: string
 *                 format: email
 *                 description: Correo del instructor a cargo.
 *                 example: instructor@empresa.com
 *               grupo:
 *                 type: string
 *                 description: Id del equipo al que se asigna el curso.
 *                 example: grp-uc-a
 *               status:
 *                 type: string
 *                 enum: [Abierto, Cerrado]
 *                 default: Abierto
 *               actividades:
 *                 type: array
 *                 items: { type: string }
 *                 description: Ids de actividades ya creadas.
 *     responses:
 *       201:
 *         description: Curso creado correctamente
 *       400:
 *         description: Error en los datos o validación fallida
 */
router.post("/create", verifyToken, requireRole("admin", "superadmin", "instructor"), CourseController.create);

/**
 * @swagger
 * /api/courses/me:
 *   get:
 *     summary: Obtiene los cursos del usuario autenticado
 *     tags: [Cursos]
 *     security:
 *       - bearerAuth: []   # Requiere token JWT
 *     responses:
 *       200:
 *         description: Lista de cursos asociados al usuario autenticado
 *         content:
 *           application/json:
 *             schema:
 *               type: array
 *               items:
 *                 type: object
 *                 properties:
 *                   id:
 *                     type: string
 *                     example: "abc123"
 *                   name:
 *                     type: string
 *                     example: "Introducción a la programación"
 *                   description:
 *                     type: string
 *                     example: "Curso básico de fundamentos de programación"
 *                   image:
 *                     type: string
 *                     example: "https://example.com/course-image.jpg"
 *       401:
 *         description: No autorizado. El token JWT es inválido o no fue enviado
 *       500:
 *         description: Error interno del servidor
 */
router.get("/me", verifyToken, CourseController.getMyCourses);

/**
 * @swagger
 * /api/courses/{id}/banner:
 *   post:
 *     summary: Sube o reemplaza la portada de un curso
 *     tags: [Cursos]
 *     description: >
 *       Recibe una imagen en multipart/form-data bajo el campo `banner`.
 *       Máximo 5 MB, solo JPEG, PNG, WebP o GIF. Queda en el bucket público
 *       "banners" de Supabase Storage y la portada anterior se borra.
 *     security:
 *       - bearerAuth: []
 *     parameters:
 *       - in: path
 *         name: id
 *         required: true
 *         schema: { type: string }
 *     requestBody:
 *       required: true
 *       content:
 *         multipart/form-data:
 *           schema:
 *             type: object
 *             properties:
 *               banner:
 *                 type: string
 *                 format: binary
 *     responses:
 *       200:
 *         description: Portada actualizada, devuelve bannerUrl
 *       400:
 *         description: Sin archivo, tipo no permitido o supera los 5 MB
 *       404:
 *         description: El curso no existe o es de otra empresa
 */
router.post(
  "/:id/banner",
  verifyToken,
  requireRole("admin", "superadmin", "instructor"),
  subida.banner,
  CourseController.subirBanner
);

module.exports = router;